requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
crcmod==1.7
cryptography==37.0.4
Deprecated==1.2.13
Flask==2.2.2
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-kms==2.12.0
google-cloud-trace==1.7.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.12.0
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-requests==0.33b0
opentelemetry-instrumentation-wsgi==0.33b0
opentelemetry-propagator-gcp==1.3.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
gunicorn==20.1.0
